# internal/server/model_login_change_req.go
/*
 * Varuna Docs
 *
 * PI Project
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package server

type LoginChangeReq struct {
	PrevPassword string `json:"prev-password,omitempty"`

	NewPassword string `json:"new-password,omitempty"`
}