# internal/server/model_login_req.go
/*
 * Varuna Docs
 *
 * PI Project
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package server

type LoginReq struct {
	Username string `json:"username"`

	Password string `json:"password"`
}