# internal/server/model_login_resp.go
/*
 * Varuna Docs
 *
 * PI Project
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package server

type LoginResp struct {
	Token string `json:"token"`

	RefreshToken string `json:"refresh-token,omitempty"`
}