# internal/server/model_refresh_req.go
/*
 * Varuna Docs
 *
 * PI Project
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package server

type RefreshReq struct {
	Token string `json:"token,omitempty"`
}