# internal/server/model_register_req.go
/*
 * Varuna Docs
 *
 * PI Project
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package server

type RegisterReq struct {
	Email string `json:"email,omitempty"`

	Password string `json:"password,omitempty"`

	Username string `json:"username,omitempty"`
}