# internal/server/model_share_permissions.go
/*
 * Varuna Docs
 *
 * PI Project
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package server

type SharePermissions struct {
	Roles []string `json:"roles,omitempty"`

	Users []string `json:"users,omitempty"`
}