# internal/server/model_user_public.go
/*
 * Varuna Docs
 *
 * PI Project
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package server

type UserPublic struct {
	Username string `json:"username,omitempty"`

	Roles []string `json:"roles,omitempty"`
}